Move to ray.util.debug.
