+++ python/ray/util/__init__.py
@@ -1,10 +1,15 @@
-from .named_actors import get_actor, register_actor
-from .actor_pool import ActorPool
-from . import iter
+from ray.util import iter
+from ray.util.actor_pool import ActorPool
+from ray.util.debug import log_once, disable_log_once_globally, \
+    enable_periodic_logging
+from ray.util.named_actors import get_actor, register_actor
 
 __all__ = [
     "ActorPool",
-    "iter",
+    "disable_log_once_globally",
+    "enable_periodic_logging",
     "get_actor",
+    "iter",
+    "log_once",
     "register_actor",
 ]
