+++ rllib/evaluation/sample_batch_builder.py
@@ -2,10 +2,10 @@ import collections
 import logging
 import numpy as np
 
+from ray.util.debug import log_once
 from ray.rllib.policy.sample_batch import SampleBatch, MultiAgentBatch
 from ray.rllib.utils.annotations import PublicAPI, DeveloperAPI
 from ray.rllib.utils.debug import summarize
-from ray.tune.utils.debug import log_once
 
 logger = logging.getLogger(__name__)
