+++ rllib/policy/dynamic_tf_policy.py
@@ -4,6 +4,7 @@ from collections import OrderedDict
 import logging
 import numpy as np
 
+from ray.util.debug import log_once
 from ray.rllib.policy.policy import Policy
 from ray.rllib.policy.sample_batch import SampleBatch
 from ray.rllib.policy.tf_policy import TFPolicy
@@ -11,7 +12,6 @@ from ray.rllib.models.catalog import ModelCatalog
 from ray.rllib.utils import try_import_tf, override
 from ray.rllib.utils.debug import summarize
 from ray.rllib.utils.tracking_dict import UsageTrackingDict
-from ray.tune.utils.debug import log_once
 
 tf = try_import_tf()
