+++ rllib/utils/tf_run_builder.py
@@ -2,8 +2,8 @@ import logging
 import os
 import time
 
+from ray.util.debug import log_once
 from ray.rllib.utils import try_import_tf
-from ray.tune.utils.debug import log_once
 
 tf = try_import_tf()
 logger = logging.getLogger(__name__)
